réorganisation code
