+++ .gitignore
@@ -2,3 +2,4 @@
 venv
 access_token
 *.jwt
+res/__py*
