+++ res/const.py
@@ -1,0 +1,13 @@
+import logging
+
+# colorama constants
+TERM_UNDERLINE = '\033[04m'
+TERM_RESET     = '\033[0m'
+
+# URL of FIDO2 repository
+# ---
+FIDO_URL = "https://mds2.fidoalliance.org/"
+
+# Define your log level
+# ---
+LOG_LEVEL = logging.INFO
